02-ai-agents/05-crash-course/openai-docs-practice/09-quiz-prep/pyproject.toml
[project]
name = "09-quiz-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
